Fixed errors that made scheme compiler not compile

+++ src/atof/algorithm/bhcomp.rs
@@ -3,6 +3,7 @@
 //! Compares the actual significant digits of the mantissa to the
 //! theoretical digits from `b+h`, scaled into the proper range.
 
+use std::convert::TryInto;
 use crate::float::*;
 use crate::float::convert::*;
 use crate::float::rounding::*;
@@ -59,7 +60,7 @@ pub(super) fn parse_mantissa<'a, Data>(data: Data, radix: u32, max_digits: usize
     let small_powers = Bigint::small_powers(radix);
     let count = data.mantissa_digits();
     let bits = count / integral_binary_factor(radix).as_usize();
-    let bytes = bits / Limb::BITS;
+    let bytes = bits / (Limb::BITS as usize);
 
     // Main loop
     let step = small_powers.len() - 2;
